refactor: move Appium session parsing and capability models to common package
